Part of WOR-165: step 2 — create watcher_helpers.py, extract pure functions
Move _parse_worker_usage, _POLICY_FLAGS, _read_result_flags,
check_allowed_paths_overlap, build_worker_env, build_worker_cmd,
resolve_effective_mode, _tee_worker_output, _parse_ollama_model
from watcher.py to watcher_helpers.py. watcher.py now imports all
nine symbols from watcher_helpers; public API is unchanged.
